Improve user API - add admin users

+++ app/routes/user.py
@@ -5,7 +5,7 @@ from ..util.auth import token_required
 
 user_bp = Blueprint('user', __name__)
 
-@user_bp.route('/user', methods=['POST'])
+@user_bp.route('/users', methods=['POST'])
 def create_user():
     data = request.get_json()
     username = data.get('username')
@@ -24,19 +24,9 @@ def create_user():
 
     return jsonify({'data': new_user.get_json()}), 201
 
-@user_bp.route('/user/<int:user_id>', methods=['GET'])
+@user_bp.route('/users/<int:user_id>', methods=['GET'])
 def get_user(user_id):
     user = db.session.get(User, user_id)
-
-    if not user:
-        return jsonify({'message': 'User not found'}), 404
-
-    return jsonify({'data': user.get_json()}), 200
-
-@user_bp.route('/user', methods=['GET'])
-@token_required
-def get_own_user(jwt_data):
-    user = db.session.get(User, jwt_data['user_id'])
 
     if not user:
         return jsonify({'message': 'User not found'}), 404
@@ -65,20 +55,37 @@ def search_users():
     result = [user.get_json() for user in users]
     return jsonify({'data': result}), 200
 
-@user_bp.route('/user', methods=['PUT'])
+@user_bp.route('/users/<int:user_id>', methods=['PUT'])
 @token_required
-def update_user(jwt_data):
-    user = db.session.get(User, jwt_data['user_id'])
+def update_user(jwt_data, user_id):
+    user = db.session.get(User, user_id)
 
     if not user:
         return jsonify({'message': 'User not found'}), 404
 
+    request_user = db.session.get(User, jwt_data['user_id'])
+
+    if not request_user:
+        return jsonify({'message': 'Request user not found'}), 404
+
+    if user.id != request_user.id and not request_user.admin:
+        return jsonify({'message': 'You do not have the right to update this user'}), 403
+
     data = request.get_json()
-    username = data.get('username', user.username)
+
+    admin = data.get('admin')
+
+    if admin:
+        if not request_user.admin:
+            return jsonify({'message': 'You do not have the right to update the "admin" field'}), 403
+
+        user.admin = admin
+
     password = data.get('password')
-    current_password = data.get('current_password')
 
     if password:
+        current_password = data.get('current_password')
+
         if not current_password:
             return jsonify({'message': 'Please provide current password'}), 400
 
@@ -87,31 +94,49 @@ def update_user(jwt_data):
 
         user.set_password(password)
 
+    username = data.get('username', user.username)
     user.username = username
     db.session.commit()
 
     return jsonify({'data': user.get_json()}), 200
 
-@user_bp.route('/user', methods=['DELETE'])
+@user_bp.route('/users/<int:user_id>', methods=['DELETE'])
 @token_required
-def delete_user(jwt_data):
-    user = db.session.get(User, jwt_data['user_id'])
+def delete_user(jwt_data, user_id):
+    user = db.session.get(User, user_id)
 
     if not user:
         return jsonify({'message': 'User not found'}), 404
+
+    request_user = db.session.get(User, jwt_data['user_id'])
+
+    if not request_user:
+        return jsonify({'message': 'Request user not found'}), 404
+
+    if user.id != request_user.id and not request_user.admin:
+        return jsonify({'message': 'You do not have the right to delete this user'}), 403
 
     db.session.delete(user)
     db.session.commit()
 
     return jsonify({'data': user.get_json()}), 200
 
-@user_bp.route('/user/decks/<int:deck_id>', methods=['POST'])
+# This actually creates a copy of the deck
+@user_bp.route('/users/<int:user_id>/decks/<int:deck_id>', methods=['POST'])
 @token_required
-def add_deck(jwt_data, deck_id):
-    user = db.session.get(User, jwt_data['user_id'])
+def add_deck(jwt_data, user_id, deck_id):
+    user = db.session.get(User, user_id)
 
     if not user:
         return jsonify({'message': 'User not found'}), 404
+
+    request_user = db.session.get(User, jwt_data['user_id'])
+
+    if not request_user:
+        return jsonify({'message': 'Request user not found'}), 404
+
+    if user.id != request_user.id and not request_user.admin:
+        return jsonify({'message': 'You do not have the right to update this user'}), 403
 
     deck = db.session.get(Deck, deck_id)
 
@@ -131,13 +156,21 @@ def add_deck(jwt_data, deck_id):
 
     return jsonify({'data': deck.get_json()}), 201
 
-@user_bp.route('/user/decks', methods=['GET'])
+@user_bp.route('/users/<int:user_id>/decks', methods=['GET'])
 @token_required
-def search_user_decks(jwt_data):
-    user = db.session.get(User, jwt_data['user_id'])
+def search_user_decks(jwt_data, user_id):
+    user = db.session.get(User, user_id)
 
     if not user:
         return jsonify({'message': 'User not found'}), 404
+
+    request_user = db.session.get(User, jwt_data['user_id'])
+
+    if not request_user:
+        return jsonify({'message': 'Request user not found'}), 404
+
+    if user.id != request_user.id and not request_user.admin:
+        return jsonify({'message': 'You do not have the right to view this user'}), 403
 
     q = request.args.get('q')
     limit = request.args.get('limit')
